move multi-image subject preloading out of flipbook's container
