+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/FlipbookViewer/FlipbookViewerContainer.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 import asyncStates from '@zooniverse/async-states'
 import PropTypes from 'prop-types'
 import { observer } from 'mobx-react'
@@ -6,7 +6,6 @@ import { observer } from 'mobx-react'
 import withKeyZoom from '@components/Classifier/components/withKeyZoom'
 import { useStores } from '@hooks'
 import locationValidator from '../../helpers/locationValidator'
-import useSubjectImage from '../SingleImageViewer/hooks/useSubjectImage'
 import FlipbookViewer from './FlipbookViewer'
 import FlipbookSeparateFrame from './components/FlipbookSeparateFrame'
 
@@ -51,39 +50,7 @@ function FlipbookViewerContainer({
     setOnPan,
     setOnZoom
   } = useStores(storeMapper)
-  /** This initializes an image element from the subject's defaultFrame src url.
-   * We do this so the SVGPanZoom has dimensions of the subject image.
-   * We're assuming all frames in one subject have the same dimensions. */
-  const defaultFrameUrl = subject
-    ? Object.values(subject.locations[defaultFrame])[0]
-    : null
-  const { img, error, loading } = useSubjectImage(defaultFrameUrl)
   const [separateFrameView, setSeparateFrameView] = useState(false)
-
-  useEffect(
-    function preloadImages() {
-      subject?.locations?.forEach(location => {
-        const [url] = Object.values(location)
-        if (url) {
-          const { Image } = window
-          const img = new Image()
-          img.src = url
-        }
-      })
-    },
-    [subject?.locations]
-  )
-
-  const { naturalHeight, naturalWidth, src: defaultFrameSrc } = img
-
-  useEffect(
-    function logError() {
-      if (!loading && error) {
-        onError(error)
-      }
-    },
-    [error, loading]
-  )
 
   const toggleSeparateFramesView = () => {
     setSeparateFrameView(!separateFrameView)
@@ -102,8 +69,6 @@ function FlipbookViewerContainer({
           key={Object.values(location)[0]}
           invert={invert}
           move={move}
-          naturalHeight={naturalHeight}
-          naturalWidth={naturalWidth}
           onError={onError}
           onKeyDown={onKeyDown}
           onReady={onReady}
@@ -119,12 +84,10 @@ function FlipbookViewerContainer({
   ) : (
     <FlipbookViewer
       defaultFrame={defaultFrame}
-      defaultFrameSrc={defaultFrameSrc}
       enableRotation={enableRotation}
       invert={invert}
       move={move}
-      naturalHeight={naturalHeight}
-      naturalWidth={naturalWidth}
+      onError={onError}
       onKeyDown={onKeyDown}
       onReady={onReady}
       playIterations={playIterations}
